LayoutTests/platform/ios-simulator/editing/pasteboard/drop-text-events-expected.txt

This tests that Drag drop fires textInput events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL actualTextEventData should be PlainHello (of type string). Was null (of type object).
FAIL testTargetTextarea.value should be PlainHello. Was .
FAIL actualTextEventData should be PlainHello (of type string). Was null (of type object).
FAIL testTargetInput.value should be PlainHello. Was .
FAIL actualTextEventData should be  (of type string). Was null (of type object).
FAIL 0 <= testTargetEditable.innerHTML.indexOf('PlainHello') should be true. Was false.
FAIL actualTextEventData should be RichHello (of type string). Was null (of type object).
FAIL testTargetTextarea.value should be RichHello. Was .
FAIL actualTextEventData should be RichHello (of type string). Was null (of type object).
FAIL testTargetInput.value should be RichHello. Was .
FAIL actualTextEventData should be  (of type string). Was null (of type object).
FAIL 0 <= testTargetEditable.innerHTML.indexOf('<b>Rich</b>Hello') should be true. Was false.
FAIL textInputCount should be 6. Was 0.
FAIL actualTextEventData should be PlainHello (of type string). Was null (of type object).
PASS testTargetTextarea.value is ''
FAIL actualTextEventData should be PlainHello (of type string). Was null (of type object).
PASS testTargetInput.value is ''
FAIL actualTextEventData should be  (of type string). Was null (of type object).
PASS testTargetEditable.innerHTML is 'placeholder'
FAIL actualTextEventData should be RichHello (of type string). Was null (of type object).
PASS testTargetTextarea.value is ''
FAIL actualTextEventData should be RichHello (of type string). Was null (of type object).
PASS testTargetInput.value is ''
FAIL actualTextEventData should be  (of type string). Was null (of type object).
PASS testTargetEditable.innerHTML is 'placeholder'
FAIL textInputCount should be 6. Was 0.
PASS successfullyParsed is true

TEST COMPLETE
